Roll up imports in products/detail/forms
