+++ src/components/products/detail/ProductDetail.tsx
@@ -13,12 +13,12 @@ import {
   Divider
 } from "@chakra-ui/react"
 import {Formik, FormikErrors, FormikTouched} from "formik"
-import DescriptionForm from "./forms/DescriptionForm"
-import DetailsForm from "./forms/DetailsForm"
+import {DescriptionForm} from "./forms/DescriptionForm/DescriptionForm"
+import {DetailsForm} from "./forms/DetailsForm/DetailsForm"
+import {InventoryForm} from "./forms/InventoryForm/InventoryForm"
+import {ShippingForm} from "./forms/ShippingForm/ShippingForm"
+import {UnitOfMeasureForm} from "./forms/UnitOfMeasureForm/UnitOfMeasureForm"
 import ImagePreview from "./ImagePreview"
-import UnitOfMeasureForm from "./forms/UnitOfMeasureForm"
-import InventoryForm from "./forms/InventoryForm"
-import ShippingForm from "./forms/ShippingForm"
 import {withDefaultValuesFallback, getObjectDiff, makeNestedObject} from "utils"
 import {cloneDeep, get, isEmpty} from "lodash"
 import {Products} from "ordercloud-javascript-sdk"
