+++ src/components/products/detail/forms/DescriptionForm/defaultValues.ts
@@ -3,8 +3,6 @@ import {ValuesType} from "types/type-helpers/ValuesType"
 
 type FieldName = ValuesType<typeof fieldNames>
 
-const defaultValues: Record<FieldName, any> = {
+export const defaultValues: Record<FieldName, any> = {
   [fieldNames.DESCRIPTION]: ""
 }
-
-export default defaultValues
