+++ src/components/products/detail/forms/DescriptionForm/index.tsx
@@ -1,9 +1,4 @@
-import {TextareaControl} from "@/components/formik"
-import {validationSchema} from "../meta"
-import * as fieldNames from "./fieldNames"
-
-type DescriptionFormProps = {}
-
-export default function DescriptionForm({}: DescriptionFormProps) {
-  return <TextareaControl name={fieldNames.DESCRIPTION} validationSchema={validationSchema} />
-}
+export * as defaultValues from "./defaultValues"
+export * as form from "./DescriptionForm"
+export * as fieldNames from "./fieldNames"
+export * as formShape from "./formShape"
