+++ src/components/products/detail/forms/DetailsForm/defaultValues.ts
@@ -3,10 +3,8 @@ import {ValuesType} from "types/type-helpers/ValuesType"
 
 type FieldName = ValuesType<typeof fieldNames>
 
-const defaultValues: Record<FieldName, any> = {
+export const defaultValues: Record<FieldName, any> = {
   [fieldNames.ACTIVE]: true,
   [fieldNames.NAME]: "",
   [fieldNames.SKU]: ""
 }
-
-export default defaultValues
