+++ src/components/products/detail/forms/DetailsForm/formShape.ts
@@ -4,10 +4,8 @@ import * as fieldNames from "./fieldNames"
 
 type FieldName = ValuesType<typeof fieldNames>
 
-const formShape: Record<FieldName, any> = {
+export const formShape: Record<FieldName, any> = {
   [fieldNames.ACTIVE]: yup.bool(),
   [fieldNames.NAME]: yup.string().max(100).required("Name is required"),
   [fieldNames.SKU]: yup.string().max(100)
 }
-
-export default formShape
