+++ src/components/products/detail/forms/InventoryForm/formShape.ts
@@ -4,11 +4,9 @@ import {ValuesType} from "types/type-helpers/ValuesType"
 
 type FieldName = ValuesType<typeof fieldNames>
 
-const formShape: Record<FieldName, any> = {
+export const formShape: Record<FieldName, any> = {
   [fieldNames.TRACK_QUANTITY]: yup.bool(),
   [fieldNames.TRACK_VARIANTS]: yup.bool(),
   [fieldNames.INVENTORY_QUANTITY_AVAILABLE]: yup.number(),
   [fieldNames.ORDER_CAN_EXCEED_INVENTORY]: yup.bool()
 }
-
-export default formShape
