+++ src/components/products/detail/forms/UnitOfMeasureForm/UnitOfMeasureForm.tsx
@@ -1,0 +1,14 @@
+import {InputControl} from "@/components/formik"
+import {validationSchema} from "../meta"
+import * as fieldNames from "./fieldNames"
+
+type UnitOfMeasureFormProps = {}
+
+export function UnitOfMeasureForm({}: UnitOfMeasureFormProps) {
+  return (
+    <>
+      <InputControl label="Quantity per unit" name={fieldNames.QUANTITY_PER_UNIT} validationSchema={validationSchema} />
+      <InputControl label="Unit of measure" name={fieldNames.UNIT_OF_MEASURE} validationSchema={validationSchema} />
+    </>
+  )
+}
